backend/ingestion-service/requirements.txt
# Libra Ingestion Sidecar — Python dependencies
# Install with: pip install -r requirements.txt

# HTTP client (for /extract-from-drive — downloads files directly from Google Drive)
requests>=2.32.0

# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.12

# PDF extraction (no external binaries required)
pymupdf>=1.24.0

# Office formats
python-docx>=1.1.0
python-pptx>=1.0.0
openpyxl>=3.1.0

# Jupyter notebooks
nbformat>=5.10.0

# HTML parsing
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Image OCR (optional — sidecar works without it, just no OCR)
# Requires Tesseract to be installed on the system:
#   Windows: https://github.com/UB-Mannheim/tesseract/wiki
#   Linux:   apt-get install tesseract-ocr
#   macOS:   brew install tesseract
pytesseract>=0.3.10
Pillow>=10.0.0
